extract error mod.
